refactor: sort by feature
Put files in folders according to feature and fix imports and exports.
