+++ index.js
@@ -1,27 +1,27 @@
-const BasicAuth = require('./administrative-sdk/models/basic-auth');
-const BasicAuthController = require('./administrative-sdk/controllers/basic-auth-controller');
-const ChoiceChallenge = require('./administrative-sdk/models/choice-challenge');
-const ChoiceChallengeController = require('./administrative-sdk/controllers/choice-challenge-controller');
-const ChoiceRecognition = require('./administrative-sdk/models/choice-recognition');
-const ChoiceRecognitionController = require('./administrative-sdk/controllers/choice-recognition-controller');
-const Connection = require('./administrative-sdk/controllers/connection-controller');
-const Organisation = require('./administrative-sdk/models/organisation');
-const OrganisationController = require('./administrative-sdk/controllers/organisation-controller');
-const Phoneme = require('./administrative-sdk/models/phoneme');
-const PronunciationAnalysis = require('./administrative-sdk/models/pronunciation-analysis');
-const PronunciationAnalysisController = require('./administrative-sdk/controllers/pronunciation-analysis-controller');
-const PronunciationChallenge = require('./administrative-sdk/models/pronunciation-challenge');
-const PronunciationChallengeController = require('./administrative-sdk/controllers/pronunciation-challenge-controller');
-const SpeechChallenge = require('./administrative-sdk/models/speech-challenge');
-const SpeechChallengeController = require('./administrative-sdk/controllers/speech-challenge-controller');
-const SpeechRecording = require('./administrative-sdk/models/speech-recording');
-const SpeechRecordingController = require('./administrative-sdk/controllers/speech-recording-controller');
-const Student = require('./administrative-sdk/models/student');
-const StudentController = require('./administrative-sdk/controllers/student-controller');
-const Tenant = require('./administrative-sdk/models/tenant');
-const TenantController = require('./administrative-sdk/controllers/tenant-controller');
-const Word = require('./administrative-sdk/models/word');
-const WordChunk = require('./administrative-sdk/models/word-chunk');
+const BasicAuth = require('./administrative-sdk/basic-auth/basic-auth');
+const BasicAuthController = require('./administrative-sdk/basic-auth/basic-auth-controller');
+const ChoiceChallenge = require('./administrative-sdk/choice-challenge/choice-challenge');
+const ChoiceChallengeController = require('./administrative-sdk/choice-challenge/choice-challenge-controller');
+const ChoiceRecognition = require('./administrative-sdk/choice-recognition/choice-recognition');
+const ChoiceRecognitionController = require('./administrative-sdk/choice-recognition/choice-recognition-controller');
+const Connection = require('./administrative-sdk/connection/connection-controller');
+const Organisation = require('./administrative-sdk/organisation/organisation');
+const OrganisationController = require('./administrative-sdk/organisation/organisation-controller');
+const Phoneme = require('./administrative-sdk/phoneme/phoneme');
+const PronunciationAnalysis = require('./administrative-sdk/pronunciation-analysis/pronunciation-analysis');
+const PronAnalaController = require('./administrative-sdk/pronunciation-analysis/pronunciation-analysis-controller');
+const PronunciationChallenge = require('./administrative-sdk/pronunciation-challenge/pronunciation-challenge');
+const PronChallController = require('./administrative-sdk/pronunciation-challenge/pronunciation-challenge-controller');
+const SpeechChallenge = require('./administrative-sdk/speech-challenge/speech-challenge');
+const SpeechChallengeController = require('./administrative-sdk/speech-challenge/speech-challenge-controller');
+const SpeechRecording = require('./administrative-sdk/speech-recording/speech-recording');
+const SpeechRecordingController = require('./administrative-sdk/speech-recording/speech-recording-controller');
+const Student = require('./administrative-sdk/student/student');
+const StudentController = require('./administrative-sdk/student/student-controller');
+const Tenant = require('./administrative-sdk/tenant/tenant');
+const TenantController = require('./administrative-sdk/tenant/tenant-controller');
+const Word = require('./administrative-sdk/word/word');
+const WordChunk = require('./administrative-sdk/word-chunk/word-chunk');
 const audioSdk = require('./audio-sdk');
 const AudioTools = require('./audio-tools');
 const CordovaMediaPlayer = require('./cordova-media-player');
@@ -52,8 +52,8 @@ module.exports = {
   ChoiceChallengeController,
   ChoiceRecognitionController,
   OrganisationController,
-  PronunciationAnalysisController,
-  PronunciationChallengeController,
+  PronunciationAnalysisController: PronAnalaController,
+  PronunciationChallengeController: PronChallController,
   SpeechChallengeController,
   SpeechRecordingController,
   StudentController,
