+++ test/speech-challengeSpec.js
@@ -1,7 +1,7 @@
 require('jasmine-ajax');
-const SpeechChallenge = require('../administrative-sdk/models/speech-challenge');
-const SpeechChallengeController = require('../administrative-sdk/controllers/speech-challenge-controller');
-const Connection = require('../administrative-sdk/controllers/connection-controller');
+const SpeechChallenge = require('../administrative-sdk/speech-challenge/speech-challenge');
+const SpeechChallengeController = require('../administrative-sdk/speech-challenge/speech-challenge-controller');
+const Connection = require('../administrative-sdk/connection/connection-controller');
 
 describe('SpeechChallenge object test', () => {
   it('should require all required fields in constructor', () => {
