+++ test/studentSpec.js
@@ -1,7 +1,7 @@
 require('jasmine-ajax');
-const Student = require('../administrative-sdk/models/student');
-const StudentController = require('../administrative-sdk/controllers/student-controller');
-const Connection = require('../administrative-sdk/controllers/connection-controller');
+const Student = require('../administrative-sdk/student/student');
+const StudentController = require('../administrative-sdk/student/student-controller');
+const Connection = require('../administrative-sdk/connection/connection-controller');
 
 describe('Student object test', () => {
   it('should instantiate a Student without id', () => {
